+++ pkg/backend/httpstate/state.go
@@ -415,6 +415,21 @@ func convertEngineEvent(e engine.Event) (apitype.EngineEvent, error) {
 			Ephemeral: p.Ephemeral,
 		}
 
+	case engine.PolicyViolationEvent:
+		p, ok := e.Payload.(engine.PolicyViolationEventPayload)
+		if !ok {
+			return apiEvent, eventTypePayloadMismatch
+		}
+		apiEvent.PolicyEvent = &apitype.PolicyEvent{
+			ResourceURN:       string(p.ResourceURN),
+			Message:           p.Message,
+			Color:             string(p.Color),
+			PolicyName:        p.PolicyName,
+			PolicyPackName:    p.PolicyPackName,
+			PolicyPackVersion: p.PolicyPackVersion,
+			EnforcementLevel:  string(p.EnforcementLevel),
+		}
+
 	case engine.PreludeEvent:
 		p, ok := e.Payload.(engine.PreludeEventPayload)
 		if !ok {